+++ ca/django_ca/tests/tests_command_dump_crl.py
@@ -138,7 +138,10 @@ class DumpCRLTestCase(DjangoCAWithCertTestCase):
             self.assertIsInstance(crl.signature_hash_algorithm, hashes.SHA512)
             self.assertEqual(len(list(crl)), 1)
             self.assertEqual(crl[0].serial_number, cert.x509.serial_number)
-            self.assertEqual(crl[0].extensions[0].value.reason.name, reason)
+
+            # unspecified is not included (see RFC 5280, 5.3.1)
+            if reason != 'unspecified':
+                self.assertEqual(crl[0].extensions[0].value.reason.name, reason)
 
     @override_tmpcadir()
     def test_compromised(self):